requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
crispy-tailwind==0.5.0
dj-database-url==2.0.0
Django==3.2.18
django-crispy-forms==2.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.5.0
whitenoise==6.4.0
gunicorn==20.1.0
django-tables2==2.5.3
loguru==0.7.0
python-dotenv
django-ninja-0.21.0
